added validators, and provided an alt impl for ldap fetching
