+++ cas-server-support-x509/pom.xml
@@ -106,6 +106,12 @@
       <version>${ldaptive.version}</version>
       <scope>test</scope>
     </dependency>
+
+    <dependency>
+      <groupId>org.hibernate</groupId>
+      <artifactId>hibernate-validator</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
   <properties>
