+++ cas-server-support-x509/src/test/resources/x509-ldap-context.xml
@@ -87,4 +87,7 @@
           p:validateOnCheckOut="${ldap.pool.validateOnCheckout}"
           p:validatePeriodically="${ldap.pool.validatePeriodically}"
           p:validatePeriod="${ldap.pool.validatePeriod}" />
+
+    <bean id="validationAnnotationBeanPostProcessor" class="org.jasig.cas.util.CustomBeanValidationPostProcessor"
+              p:afterInitialization="true"/>
 </beans>
